requirements.txt
annotated-types==0.6.0
anyio==4.2.0
appdirs==1.4.4
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
diskcache==5.6.3
distro==1.9.0
FLAML==2.1.1
frozendict==2.4.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.2
httpx==0.26.0
idna==3.6
lxml==5.0.0
multitasking==0.0.11
numpy==1.26.3
openai==1.6.1
pandas==2.1.4
peewee==3.17.0
pyautogen==0.2.2
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
regex==2023.12.25
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
termcolor==2.4.0
tiktoken==0.5.2
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
webencodings==0.5.1
yfinance==0.2.33
